# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_application_collaboration.DeleteNodeResult:
    properties:
      bodyEpoch:
        type: integer
      bodyVersion:
        type: integer
      changed:
        type: boolean
      commandID:
        type: string
      documentID:
        type: string
      nodeID:
        type: string
    type: object
  backend_internal_application_collaboration.MoveNodeResult:
    properties:
      bodyEpoch:
        type: integer
      bodyVersion:
        type: integer
      changed:
        type: boolean
      commandID:
        type: string
      documentID:
        type: string
    type: object
...
  backend_internal_domain_model.DocumentAccess:
    properties:
      accessLevel:
        description: owner, edit, comment, view
        type: string
      createdAt:
        type: string
      documentId:
        type: string
      user:
        $ref: '#/definitions/backend_internal_domain_model.UserAuthor'
      userId:
        type: string
    type: object
...
  backend_internal_domain_model.UserAuthor:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  backend_internal_presentation_document_presenter.AddAccessRequest:
    properties:
      email:
        type: string
      level:
        type: string
    type: object
...
  backend_internal_presentation_response.Envelope:
    properties:
      data: {}
    type: object
  backend_internal_presentation_response.ErrorEnvelope:
    properties:
      title:
        type: string
    type: object
...
  internal_presentation_document_handler.bodyInitializationNode:
    properties:
      attributes:
        items:
          type: integer
        type: array
      content:
        type: string
      nodeID:
        type: string
      parentID:
        type: string
      siblingOrder:
        type: number
      type:
        type: string
    type: object
  internal_presentation_document_handler.bodyInitializationRequest:
    properties:
      baseBodyVersion:
        type: integer
      bodySchemaVersion:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/internal_presentation_document_handler.bodyInitializationNode'
        type: array
      rootNodeID:
        type: string
      sourceFingerprint:
        type: string
    type: object
  internal_presentation_document_handler.bodyNodeResponse:
    properties:
      attributes:
        items:
          type: integer
        type: array
      content:
        type: string
      nodeID:
        type: string
      parentID:
        type: string
      siblingOrder:
        type: number
      type:
        type: string
      version:
        type: integer
    type: object
  internal_presentation_document_handler.bodySnapshotResponse:
    properties:
      bodyEpoch:
        type: integer
      bodySchemaVersion:
        type: integer
      bodyVersion:
        type: integer
      canEdit:
        type: boolean
      encodedState:
        items:
          type: integer
        type: array
      nodes:
        items:
          $ref: '#/definitions/internal_presentation_document_handler.bodyNodeResponse'
        type: array
      rootNodeID:
        type: string
    type: object
...
  internal_presentation_document_handler.deleteNodeRequest:
    properties:
      bodyEpoch:
        type: integer
      bodySchemaVersion:
        type: integer
      commandID:
        type: string
      nodeID:
        type: string
    type: object
  internal_presentation_document_handler.moveNodeRequest:
    properties:
      beforeNodeID:
        type: string
      bodyEpoch:
        type: integer
      bodySchemaVersion:
        type: integer
      commandID:
        type: string
      nodeID:
        type: string
      targetParentID:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for Dokudocs backend service.
  title: Dokudocs API
  version: "1.0"
paths:
...
  /documents/{id}/accesses:
    get:
      description: Get list of collaborators and their permissions for a document
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_model.DocumentAccess'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List document accesses
      tags:
      - Document
    post:
      consumes:
      - application/json
      description: Grant document access to a user by email
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Add access payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_document_presenter.AddAccessRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add or update document access
      tags:
      - Document
...
  /documents/{id}/body:
    get:
      description: Returns the canonical AST and matching Yjs state for a Markdown
        document.
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_presentation_document_handler.bodySnapshotResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Markdown body
      tags:
      - Document
  /documents/{id}/body/delete:
    post:
      consumes:
      - application/json
      description: Atomically deletes a block subtree and advances the document body
        epoch.
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete command
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_presentation_document_handler.deleteNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_application_collaboration.DeleteNodeResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete a Markdown block
      tags:
      - Document
  /documents/{id}/body/initialize:
    post:
      consumes:
      - application/json
      description: Atomically stores the AST and initial Yjs state for an uninitialized
        Markdown document.
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Parsed body and source fingerprint
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_presentation_document_handler.bodyInitializationRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Initialize Markdown body
      tags:
      - Document
  /documents/{id}/body/move:
    post:
      consumes:
      - application/json
      description: Applies one idempotent structural move and returns the resulting
        body epoch/version.
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Move command
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_presentation_document_handler.moveNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_application_collaboration.MoveNodeResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Move a Markdown block
      tags:
      - Document
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"